🚚 api: [PRESCRIPTION] move service to src/prescription/campaign context
