+++ api/src/prescription/campaign/domain/services/campaign-csv-export-service.js
@@ -1,6 +1,6 @@
 import * as campaignParticipationService from '../../../../../lib/domain/services/campaign-participation-service.js';
 import * as csvSerializer from '../../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
-import { CampaignAssessmentCsvLine } from '../../../../../lib/infrastructure/utils/CampaignAssessmentCsvLine.js';
+import { CampaignAssessmentCsvLine } from '../../infrastructure/utils/CampaignAssessmentCsvLine.js';
 
 export { createOneCsvLine };
